펫 몸무게 소수처리

+++ Back_toMyAnimal/src/main/java/team1/toMyAnimal/domain/dto/pet/PetCreateRequest.java
@@ -30,7 +30,7 @@ public class PetCreateRequest {
     private String birthday;
 
 
-    private Long weight;
+    private double weight;
 
     @Null
     private Long memberId;
